Fix lots and lots of warnings
Fixes a bunch of warnings and brings the code base more inline (style wise) with `Ably.NET`.

I have applied things like `readonly` where possible and used `var` where the inferred type is obvious from the right-hand side of the expression.

I have also normalised the mixed tabs and spaces to the defacto standard of 4 spaces for a logical tab.

If you review this in an external diff tool and enable Ignore All Whitespace Changes you'll have a much easier time reviewing it.